// server/src/common/exceptions/domain-error.enum.ts
export enum DomainErrorCode {
  SESSION_NOT_FOUND = 'SESSION_NOT_FOUND',
  INVALID_GAME = 'INVALID_GAME',
  INVALID_GAME_STATE = 'INVALID_GAME_STATE',
  GAME_ALREADY_SUBMITTED = 'GAME_ALREADY_SUBMITTED',
  INVALID_GAME_RESULT = 'INVALID_GAME_RESULT',
  PHOTO_REQUIRED = 'PHOTO_REQUIRED',
  UPLOAD_INVALID = 'UPLOAD_INVALID',
  NOT_ALL_GAMES_COMPLETED = 'NOT_ALL_GAMES_COMPLETED',
  STARPRINT_ALREADY_GENERATED = 'STARPRINT_ALREADY_GENERATED',
  STARPRINT_NOT_FOUND = 'STARPRINT_NOT_FOUND',
  UNAUTHORIZED_SESSION = 'UNAUTHORIZED_SESSION',
  UNAUTHORIZED_MUTATION = 'UNAUTHORIZED_MUTATION',
  CONSENT_INVALID = 'CONSENT_INVALID',
  // Activities
  NEWS_NOT_FOUND = 'NEWS_NOT_FOUND',
  EVENT_NOT_FOUND = 'EVENT_NOT_FOUND',
  EVENT_REGISTRATION_CLOSED = 'EVENT_REGISTRATION_CLOSED',
  EVENT_FULL = 'EVENT_FULL',
  DUPLICATE_REGISTRATION = 'DUPLICATE_REGISTRATION',
}